test: enable FindReplace-test
